Refactor coordinator module and ensure cache invalidation on coordinator->worker stream drop
